Move plugin code to plugin.c
