+++ src/Makefile.in
@@ -26,11 +26,11 @@ C_FILES=access.c atr_tab.c attrib.c boolexp.c bsd.c bufferq.c		\
 	help.c htab.c intmap.c local.c lock.c log.c look.c malias.c	\
 	map_file.c markup.c match.c memcheck.c move.c mycrypt.c		\
 	mymalloc.c mysocket.c myrlimit.c myssl.c notify.c parse.c	\
-	pcg_basic.c player.c plyrlist.c predicat.c privtab.c		\
-	info_master.c ptab.c rob.c services.c set.c sig.c sort.c	\
-	speech.c spellfix.c sql.c sqlite3.c ssl_master.c strdup.c	\
-	strtree.c strutil.c tables.c testframework.c timer.c tz.c	\
-	uint.c unparse.c utf_impl.c utils.c version.c wait.c		\
+	pcg_basic.c player.c plugin.c plyrlist.c predicat.c 		\
+	privtab.c info_master.c ptab.c rob.c services.c set.c sig.c	\
+	sort.c speech.c spellfix.c sql.c sqlite3.c ssl_master.c 	\
+	strdup.c strtree.c strutil.c tables.c testframework.c timer.c 	\
+	tz.c uint.c unparse.c utf_impl.c utils.c version.c wait.c	\
 	warnings.c websock.c wild.c wiz.c
 
 
@@ -44,7 +44,7 @@ O_FILES=access.o atr_tab.o attrib.o boolexp.o bsd.o bufferq.o		\
 	help.o htab.o intmap.o local.o lock.o log.o look.o malias.o	\
 	map_file.o markup.o match.o memcheck.o move.o mycrypt.o		\
 	mymalloc.o mysocket.o myrlimit.o myssl.o notify.o parse.o	\
-	pcg_basic.o player.o plyrlist.o predicat.o privtab.o		\
+	pcg_basic.o player.o plugin.o plyrlist.o predicat.o privtab.o \
 	info_master.o ptab.o rob.o services.o set.o sig.o sort.o	\
 	speech.o spellfix.o sql.o sqlite3.o ssl_master.o strdup.o	\
 	strtree.o strutil.o tables.o testframework.o timer.o tz.o	\
@@ -1868,6 +1868,11 @@ player.o: ../hdrs/parse.h
 player.o: ../hdrs/mushsql.h
 player.o: ../hdrs/sqlite3.h
 player.o: ../hdrs/strutil.h
+plugin.o: ../hdrs/copyrite.h
+plugin.o: ../hdrs/conf.h
+plugin.o: ../hdrs/log.h
+plugin.o: ../hdrs/mymalloc.h
+plugin.o: ../hdrs/plugin.h
 plyrlist.o: ../config.h
 plyrlist.o: ../confmagic.h
 plyrlist.o: ../options.h
